requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
anthropic==0.39.0
supabase==2.3.4
python-dotenv==1.0.0
flask==3.0.0
schedule==1.2.0
